separate gpu code from flatlinearindexstring to gpulinearindex.
