+++ app/benchmark.ts
@@ -7,11 +7,12 @@ import {
     HybridBigramInvertedIndex,
     LinearIndex,
     LinearIndexString,
+    GPULinearIndex,
     UniSearchError,
     createIndex,
     search,
 } from "@src/main";
-import { wikipedia_ja_extracted } from "@test/wikipedia_ja_extracted";
+import { wikipedia_ja_extracted_1000 } from "@test/wikipedia_ja_extracted_1000";
 import { wikipedia_ja_keyword } from "@test/wikipedia_ja_keyword";
 import { calculateGzipedJsonSize } from "@ref/util";
 type BenchmarkResult = {
@@ -92,13 +93,15 @@ export async function execBenchmark(
     return benchmark_results;
 }
 
-const keywords = getAllKeywords(wikipedia_ja_keyword).slice(0, 50);
-const benchmark_results_linear = await execBenchmark(LinearIndex, {}, wikipedia_ja_extracted, keywords);
-const benchmark_results_linear_string = await execBenchmark(LinearIndexString, {}, wikipedia_ja_extracted, keywords);
-const benchmark_results_flat = await execBenchmark(FlatLinearIndex, {}, wikipedia_ja_extracted, keywords);
-const benchmark_results_flat_string = await execBenchmark(FlatLinearIndexString, {}, wikipedia_ja_extracted, keywords);
-const benchmark_results_flat_string_delegated = await execBenchmark(FlatLinearIndexStringDelegated, {}, wikipedia_ja_extracted, keywords);
-const benchmark_results_hybrid = await execBenchmark(HybridBigramInvertedIndex, {}, wikipedia_ja_extracted, keywords);
+const keywords = getAllKeywords(wikipedia_ja_keyword).slice(0, 30);
+const articles = wikipedia_ja_extracted_1000.slice(0, 1000);
+const benchmark_results_linear = await execBenchmark(LinearIndex, {}, articles, keywords);
+const benchmark_results_linear_string = await execBenchmark(LinearIndexString, {}, articles, keywords);
+const benchmark_results_flat = await execBenchmark(FlatLinearIndex, {}, articles, keywords);
+const benchmark_results_flat_string = await execBenchmark(FlatLinearIndexString, {}, articles, keywords);
+const benchmark_results_flat_string_delegated = await execBenchmark(FlatLinearIndexStringDelegated, {}, articles, keywords);
+const benchmark_results_gpu = await execBenchmark(GPULinearIndex, {}, articles, keywords);
+const benchmark_results_hybrid = await execBenchmark(HybridBigramInvertedIndex, {}, articles, keywords);
 
 console.log("benchmark results: LinearIndex");
 console.log(benchmark_results_linear);
@@ -110,5 +113,7 @@ console.log("benchmark results: FlatLinearIndexString");
 console.log(benchmark_results_flat_string);
 console.log("benchmark results: FlatLinearIndexStringDelegated");
 console.log(benchmark_results_flat_string_delegated);
+console.log("benchmark results: GPULinearIndex");
+console.log(benchmark_results_gpu);
 console.log("benchmark results: HybridBigramInvertedIndex");
 console.log(benchmark_results_hybrid);
