+++ src/frontend/indextypes.ts
@@ -6,6 +6,7 @@ import { Hybrid } from "@src/method/hybrid";
 import { InvertedIndex } from "@src/method/invertedindex";
 import { LinearIndex } from "@src/method/linearindex";
 import { LinearIndexString } from "@src/method/linearindexString";
+import { GPULinearIndex } from "@src/method/gpulinearindex";
 import { Ngram } from "@src/method/ngram";
 
 export const HybridBigramInvertedIndex = Hybrid("HybridBigramInvertedIndex", Ngram(2, InvertedIndex), InvertedIndex);
@@ -17,4 +18,5 @@ export const IndexTypes: { [key: string]: IndexClass } = {
     FlatLinearIndexString,
     FlatLinearIndexStringDelegated,
     LinearIndexString,
+    GPULinearIndex,
 };
