+++ src/main.ts
@@ -6,6 +6,7 @@ export { FlatLinearIndex } from "@src/method/flatlinearindex";
 export { FlatLinearIndexString } from "@src/method/flatlinearindexstring";
 export { FlatLinearIndexStringDelegated } from "@src/method/flatlinearindexstringdelegated";
 export { LinearIndexString } from "@src/method/linearindexString";
+export { GPULinearIndex } from "@src/method/gpulinearindex";
 export { UniSearchError } from "@src/frontend/base";
 export type {
     UniIndex,
